Extract safe socks5 helpers into focused files
